SQL - ændret bigint i balance til double precision
Tilføjet User klasse med to constructor

Tilføjet UserController
- addRoutes
- registerUser
- login
- logout

+++ src/main/resources/sql/cupcake_db.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS public.users
     user_id serial NOT NULL,
     username character varying,
     password character varying,
-    balance bigint,
+    balance double precision,
     administrator boolean,
     PRIMARY KEY (user_id)
     );
